refactor imports to use absolute paths for better readability and maintainability
