+++ src/app/(application)/(tabs)/home.tsx
@@ -1,4 +1,4 @@
-import { HomeScreen } from "../../../ui/screens/Home/Home"
+import { HomeScreen } from "@/screens"
 
 export default function Home() {
 	return <HomeScreen />
