+++ src/app/index.tsx
@@ -1,6 +1,6 @@
 import { Redirect } from "expo-router"
 import { useAuth } from "../domain/auth/AuthContext"
-import { OnboardingScreen } from "../ui/screens/Onboarding/Onboarding"
+import { OnboardingScreen } from "@/screens"
 
 export default function Index() {
 	const { auth } = useAuth()
