+++ src/ui/components/containers/index.ts
@@ -1,0 +1,1 @@
+export * from "./TabBar"
