+++ src/ui/components/core/Button/ButtonContent.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from "react"
-import { Text } from "../Text/Text"
+import { Text } from "@/components/core"
 import { useButtonContext } from "./ButtonContext"
 
 export const ButtonContent = ({ children }: PropsWithChildren) => {
