+++ src/ui/components/core/Header/HeaderVerticalCenterTitle.tsx
@@ -1,4 +1,4 @@
-import { Text } from "../Text/Text"
+import { Text } from "@/components/core"
 import { Header } from "./HeaderTypes"
 import { styles } from "./styles"
