+++ src/ui/components/core/Screen/index.ts
@@ -1,0 +1,2 @@
+export * from "./Screen"
+export { type Screen as ScreenTypes } from "./ScreenTypes"
