+++ src/ui/components/core/index.ts
@@ -1,0 +1,6 @@
+export * from "./Button"
+export * from "./Header"
+export * from "./Icon"
+export * from "./Pressable"
+export * from "./Screen"
+export * from "./Text"
