+++ src/ui/screens/Onboarding/__tests__/Onboarding.test.tsx
@@ -1,4 +1,4 @@
-import { render } from "../../../../infra/tests"
+import { render } from "@/tests"
 import { OnboardingScreen } from "../Onboarding"
 
 describe("Onboarding Screen", () => {
