+++ src/ui/screens/Onboarding/index.ts
@@ -1,0 +1,1 @@
+export * from "./Onboarding"
