+++ src/ui/screens/WorkoutList/index.ts
@@ -1,0 +1,1 @@
+export * from "./WorkoutList"
