+++ src/ui/theme/index.ts
@@ -1,2 +1,4 @@
-export * from "./tokens/colors/index"
+export * from "./tokens"
 export * from "./types"
+export * from "./hooks"
+export * from "./provider"
