+++ src/ui/theme/provider/index.ts
@@ -1,0 +1,1 @@
+export * from "./themeProvider"
